src/ai/requirements.txt
google-generativeai==0.3.1
google-cloud-logging==3.8.0
